Implemented Level Switching
Now when the player moves off the screen the level changes as well as Nova wrapping around to the opposite side. Moving left switches to the previous level, moving right switches to the next level. The levels also wrap so moving left off the first level switches to the final level in the list and vice versa.